Enhanced pause/resume with gated restarts - 2
This contains a fix against the original change id:
Ie0c5e0249bde0839345ad66f7400522754aa91ca which broke
keystone. Otherwise, the fix is the same:

The existing pause/resume functionality is enhanced with
changed charm-helpers support to chech that the services
really are stopped and that paused units really stay
paused.  The restart_on_change decorator is gated
such that if the unit is 'paused' then the service
is not accidentally started.

Change-Id: I6a828676be11338266845e822be087d734944da0