Ensure log-level config option is applied consistently.
Keystone charm currently supports 'log-level' but it is not handled
at all. This setting should be applied independently (presumably it
was the original intention for this config option) from debug
and verbose mode.

It also takes care of improper value set by the user which in case user
enters a non log-level option the default 'ERROR' is used instead.

Change-Id: I53cc0cafed7e56be9d568726f7590a36d3be1a87
Closes-Bug: #1554865
Signed-off-by: Leonardo Borda <leonardo.borda@canonical.com>