Switch to using charm-store for amulet tests
All OpenStack charms are now directly published to the charm store
on landing; switch Amulet helper to resolve charms using the
charm store rather than bzr branches, removing the lag between
charm changes landing and being available for other charms to
use for testing.

This is also important for new layered charms where the charm must
be build and published prior to being consumable.

This patch also fixes a potential restart of apache2 when the unit
is in paused state during the execution of config-changed; this is
the root cause of the intermittent amulet test failures.

Change-Id: I90e61f5c217ae709bc30ffe16b1569b9fd2c5719